--- src/net/reliable_udp/mod.rs
@@ -1,9 +1,0 @@
-mod connection;
-mod packet;
-mod udp;
-
-use self::connection::{Connection, ConnectionQuality };
-use self::packet::{ AckRecord, ExternalAcks };
-use Packet;
-
-pub use self::udp::UdpSocket;
