+++ src/net/udp.rs
@@ -1,8 +1,10 @@
-use bincode::{deserialize, serialize};
-use packet::Packet;
 use std::io;
 use std::net::{self, ToSocketAddrs, SocketAddr};
 use std::collections::HashMap;
+
+use bincode::{deserialize, serialize};
+use packet::Packet;
+use super::*;
 
 const BUFFER_SIZE: usize = 1024;
 
@@ -29,7 +31,7 @@ impl UdpSocket {
         let (len, _addr) = self.socket.recv_from(&mut self.recv_buffer)?;
 
         if len > 0 {
-            // TODO: Remove unwrap and funnel result error typesq
+            // TODO: Remove unwrap and funnel result error types
             let packet: Packet = deserialize(&self.recv_buffer[..len]).unwrap();
             self.state.process_received(_addr, &packet);
             Ok(Some(packet))
@@ -38,34 +40,15 @@ impl UdpSocket {
         }
     }
 
-    pub fn send(&mut self, packet: &Packet) -> io::Result<usize> {
-        let buf = serialize(packet).unwrap();
-        let (addr, payload) = self.state.pre_process_packet(*packet);
+    pub fn send(&mut self, packet: Packet) -> io::Result<usize> {
+        let mut packet = packet;
+        let (addr, payload) = self.state.pre_process_packet(packet);
 
         self.socket.send_to(&payload, addr)
     }
 
     pub fn set_nonblocking(&mut self, nonblocking: bool) -> io::Result<()> {
         self.socket.set_nonblocking(nonblocking)
-    }
-}
-
-pub struct Connection
-{
-    pub seq_num: u16,
-    pub dropped_packets: Vec<Packet>,
-    pub waiting_packets: AckRecord,
-    pub their_acks: ExternalAcks,
-}
-
-impl Connection {
-    pub fn new() -> Connection {
-        Connection {
-            seq_num: 0,
-            dropped_packets: Vec::new(),
-            waiting_packets: AckRecord::new(),
-            their_acks: ExternalAcks::new()
-        }
     }
 }
 
@@ -79,7 +62,7 @@ impl SocketState {
     }
 
     pub fn pre_process_packet(&mut self, packet: Packet) -> (SocketAddr, Vec<u8>) {
-        let connection = self.create_connection_if_not_exists(&addr);
+        let connection = self.create_connection_if_not_exists(&packet.addr);
 
         // queue new packet
         connection.waiting_packets.enqueue(connection.seq_num, packet.clone());
@@ -89,7 +72,11 @@ impl SocketState {
 
         // increase sequence number
         connection.seq_num = connection.seq_num.wrapping_add(1);
-        (packet.addr, final_packet.serialized())
+
+        // TODO: remove unwrap
+        let buffer = serialize(&final_packet).unwrap();
+
+        (final_packet.addr, buffer)
     }
 
     pub fn dropped_packets(&mut self, addr: SocketAddr) -> Vec<Packet> {
@@ -99,14 +86,15 @@ impl SocketState {
 
     pub fn process_received(&mut self, addr: SocketAddr, packet: &Packet) {
         let mut connection = self.create_connection_if_not_exists(&addr);
-        connection.their_acks.ack(packet.seq);
+        connection.their_acks.ack(packet.seq.unwrap());
 
         // get dropped packets
-        let dropped_packets = connection.waiting_packets.ack(packet.ack_seq, packet.ack_field);
+        let dropped_packets = connection.waiting_packets.ack(packet.ack_seq.unwrap(), packet.ack_field.unwrap());
         connection.dropped_packets = dropped_packets.into_iter().map(|(_, p)| p).collect();
     }
 
-    fn create_connection_if_not_exists(&mut self, add: &SocketAddr) -> &mut Connection
+    #[inline]
+    fn create_connection_if_not_exists(&mut self, addr: &SocketAddr) -> &mut Connection
     {
         self.connections.entry(*addr).or_insert(Connection::new())
     }
